refactor: move `mapOrgModelToGQL` into dedicated file, add missing asana links
